requirements.txt
fastapi==0.109.0
uvicorn==0.26.0
python-dotenv==1.0.0
azure-cognitiveservices-speech==1.34.1
